Create date-range-rag-days-left.json
I loved the 'date-range-rag.json' example but needed to extend it to show the remaining days. I was having trouble getting a whole number for 'days left' (the number of days to the nearest days until the 'DueDate') so I embarked on producing a (pretty ugly) nested if statement that will display the number of days until the DueDate, up to 100 days, after which it will just say 100+. 

If there is a better way  (which I'm sure there is) of getting a whole number to represent days left (rather than the 15 decimal point number which is just too accurate for me!), then I'd love to hear about it but I spent some time on this and thought it was worth sharing with the original authors of 'date-range-rag.json'. 

Thanks,
Joe