Update generic-yearly-gantt-chart-with-today.json
I hope I'm doing this right (this is my first time editing on GitHub).  I've updated the JSON so that the Gantt chart now correctly renders when Start and End dates are outside the timeframe of the current year.  That is, if the Start and/or End dates are before 1st January of current year, as well as if Start and/or End dates are after 31st December of current year.